cdk/requirements.txt
# Mantener alineado con la versión mayor del CLI de CDK (`npx aws-cdk@2.170.0` o la que uses en CI).
aws-cdk-lib==2.170.0
constructs>=10.0.0,<11
